nms_adpcm: fix int overflow in signal estimate
It is possible (though functionally incorrect) for the signal estimate
calculation in nms_adpcm_update() to overflow the int value of s_e,
resulting in undefined behavior.

NMS ADPCM signal values cannot exceed 16 bits of width, by spec. So
pre-cast the signal components to u_fast16_t to provide enough numeric
space for the calculation.

+++ src/nms_adpcm.c
@@ -258,7 +258,7 @@ nms_adpcm_update (struct nms_adpcm_state *s)
 		} ;
 
 	/* Compute the signal estimate. */
-	s->s_e = s->a [0] * s->s_r [0] + s->a [1] * s->s_r [1] + s->s_ez ;
+	s->s_e = (int_fast16_t) s->a [0] * s->s_r [0] + (int_fast16_t) s->a [1] * s->s_r [1] + (int_fast16_t) s->s_ez ;
 
 	/* Return to scale */
 	s->s_ez >>= 14 ;
